Move project commands to commands.go
