--- project/project_build.go
@@ -1,17 +1,0 @@
-package project
-
-import (
-	"golang.org/x/net/context"
-
-	"github.com/rancher/rancher-compose-executor/project/events"
-	"github.com/rancher/rancher-compose-executor/project/options"
-)
-
-// Build builds the specified services (like docker build).
-func (p *Project) Build(ctx context.Context, buildOptions options.Build, services ...string) error {
-	return p.perform(events.ProjectBuildStart, events.ProjectBuildDone, services, wrapperAction(func(wrapper *serviceWrapper, wrappers map[string]*serviceWrapper) {
-		wrapper.Do(wrappers, events.ServiceBuildStart, events.ServiceBuild, func(service Service) error {
-			return service.Build(ctx, buildOptions)
-		})
-	}), nil)
-}
