--- project/project_log.go
@@ -1,16 +1,0 @@
-package project
-
-import (
-	"golang.org/x/net/context"
-
-	"github.com/rancher/rancher-compose-executor/project/events"
-)
-
-// Log aggregates and prints out the logs for the specified services.
-func (p *Project) Log(ctx context.Context, follow bool, services ...string) error {
-	return p.forEach(services, wrapperAction(func(wrapper *serviceWrapper, wrappers map[string]*serviceWrapper) {
-		wrapper.Do(nil, events.NoEvent, events.NoEvent, func(service Service) error {
-			return service.Log(ctx, follow)
-		})
-	}), nil)
-}
